Pushing remaining changes for Vin
