+++ components/chat/Chat.jsx
@@ -1,6 +1,14 @@
 import UserList from './UserList';
+import { useEffect } from 'react';
+// Will it be protected or not?
+// import axiosWithToken from '../redux/axios';
 
 const Chat = () => {
+  //   useEffect = () => {
+  //     // Fetch messages
+  //     // Return Messages
+  //   };
+
   return (
     <div>
       <UserList></UserList>
